chore: ResourceViewItemAction types
