+++ packages/frontend/src/components/common/ResourceView/ResourceActionHandlers.tsx
@@ -1,8 +1,5 @@
 import {
     assertUnreachable,
-    ResourceViewChartItem,
-    ResourceViewDashboardItem,
-    ResourceViewItem,
     ResourceViewItemType,
     Space,
 } from '@lightdash/common';
@@ -31,49 +28,7 @@ import DashboardDeleteModal from '../modal/DashboardDeleteModal';
 import DashboardUpdateModal from '../modal/DashboardUpdateModal';
 import SpaceActionModal from '../SpaceActionModal';
 import { ActionType } from '../SpaceActionModal/types';
-
-export enum ResourceViewItemAction {
-    CLOSE,
-    UPDATE,
-    DELETE,
-    DUPLICATE,
-    ADD_TO_DASHBOARD,
-    CREATE_SPACE,
-    MOVE_TO_SPACE,
-    PIN_TO_HOMEPAGE,
-}
-
-export type ResourceViewItemActionState =
-    | { type: ResourceViewItemAction.CLOSE }
-    | {
-          type: ResourceViewItemAction.UPDATE;
-          item: ResourceViewItem;
-      }
-    | {
-          type: ResourceViewItemAction.DELETE;
-          item: ResourceViewItem;
-      }
-    | {
-          type: ResourceViewItemAction.DUPLICATE;
-          item: ResourceViewChartItem | ResourceViewDashboardItem;
-      }
-    | {
-          type: ResourceViewItemAction.ADD_TO_DASHBOARD;
-          item: ResourceViewChartItem;
-      }
-    | {
-          type: ResourceViewItemAction.CREATE_SPACE;
-          item: ResourceViewChartItem | ResourceViewDashboardItem;
-      }
-    | {
-          type: ResourceViewItemAction.PIN_TO_HOMEPAGE;
-          item: ResourceViewItem;
-      }
-    | {
-          type: ResourceViewItemAction.MOVE_TO_SPACE;
-          item: ResourceViewChartItem | ResourceViewDashboardItem;
-          data: { spaceUuid: string };
-      };
+import { ResourceViewItemAction, ResourceViewItemActionState } from './types';
 
 interface ResourceActionHandlersProps {
     action: ResourceViewItemActionState;
