+++ packages/frontend/src/components/common/ResourceView/types.ts
@@ -1,0 +1,48 @@
+import {
+    ResourceViewChartItem,
+    ResourceViewDashboardItem,
+    ResourceViewItem,
+} from '@lightdash/common';
+
+export enum ResourceViewItemAction {
+    CLOSE,
+    UPDATE,
+    DELETE,
+    DUPLICATE,
+    ADD_TO_DASHBOARD,
+    CREATE_SPACE,
+    MOVE_TO_SPACE,
+    PIN_TO_HOMEPAGE,
+}
+
+export type ResourceViewItemActionState =
+    | { type: ResourceViewItemAction.CLOSE }
+    | {
+          type: ResourceViewItemAction.UPDATE;
+          item: ResourceViewItem;
+      }
+    | {
+          type: ResourceViewItemAction.DELETE;
+          item: ResourceViewItem;
+      }
+    | {
+          type: ResourceViewItemAction.DUPLICATE;
+          item: ResourceViewChartItem | ResourceViewDashboardItem;
+      }
+    | {
+          type: ResourceViewItemAction.ADD_TO_DASHBOARD;
+          item: ResourceViewChartItem;
+      }
+    | {
+          type: ResourceViewItemAction.CREATE_SPACE;
+          item: ResourceViewChartItem | ResourceViewDashboardItem;
+      }
+    | {
+          type: ResourceViewItemAction.PIN_TO_HOMEPAGE;
+          item: ResourceViewItem;
+      }
+    | {
+          type: ResourceViewItemAction.MOVE_TO_SPACE;
+          item: ResourceViewChartItem | ResourceViewDashboardItem;
+          data: { spaceUuid: string };
+      };
